backend/requirements.txt
fastapi
uvicorn
robin_stocks
pydantic
pydantic-settings
python-dotenv
openai
yfinance>=0.2.36
finnhub-python>=2.4.18
requests
pandas
ta>=0.11.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
sse-starlette>=1.8.0
